requirements.txt
Flask==2.3.3
flask-pymongo==2.3.0
flask-cors==4.0.0
bcrypt==4.0.1
python-dotenv==1.0.0
pymongo==4.5.0